requirements.txt
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
supabase==2.0.3
python-dotenv==1.0.0
requests==2.31.0